+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ------------------
+...
+
+1.0 (2018-03-07)
+------------------
 #. Add Wagtail 2.0 compatibility.
 #. Ensure captcha field is properly removed from submission data. (issue #11)
 #. Allow subclassing of the form builder. (issue #7)